Initialization gold as 0 in Equipment constructor
I forgot to initialize a value of gold in Equipment class. Now it is
zero for every Equipment object but I will change it when I will working
on Creatures and their equipment - I will have to declare that some
beings have money, some not. But first I have to create basic classes.